changed to create cell from nib
