+++ AlcatrazTour/PluginListMainViewController.swift
@@ -1,5 +1,5 @@
 //
-//  ViewController.swift
+//  PluginListMainViewController
 //  AlcatrazTour
 //
 //  Copyright (c) 2015年 haranicle. All rights reserved.
@@ -7,45 +7,17 @@
 
 import UIKit
 import Realm
-import M2DWebViewController
 
-enum Modes:Int {
-    case Popularity = 0
-    case Stars = 1
-    case Update = 2
-    case New = 3
-    
-    func toIcon() -> String {
-        switch self {
-            case Modes.Popularity: return "\u{f004}"
-            case Modes.Stars: return "\u{f005}"
-            case Modes.Update: return "\u{f021}"
-            case Modes.New: return "\u{f135}"
-            default: return ""
-        }
-    }
-    
-    func toString() -> String {
-        switch self {
-        case Modes.Popularity: return "Popularity"
-        case Modes.Stars: return "Stars"
-        case Modes.Update: return "Update"
-        case Modes.New: return "New"
-        default: return ""
-        }
-    }
-}
-
-class ViewController: UIViewController {
+class PluginListMainViewController: PluginListBaseViewController {
     
     var githubClient = GithubClient()
     var currentMode = Modes.Popularity
     let segments = [Modes.Popularity, Modes.Stars, Modes.Update, Modes.New]
-
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        // notification center 
+        // notification center
         NSNotificationCenter.defaultCenter().addObserver(self, selector: "onApplicationDidBecomeActive:", name: UIApplicationDidBecomeActiveNotification, object: nil)
         
         let attributes = [NSFontAttributeName:UIFont(name: "FontAwesome", size: 10)!]
@@ -59,14 +31,6 @@ class ViewController: UIViewController {
     
     deinit{
         NSNotificationCenter.defaultCenter().removeObserver(self)
-    }
-    
-    override func viewWillAppear(animated: Bool) {
-        navigationController!.toolbarHidden = true
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
     }
     
     func onApplicationDidBecomeActive(notification:NSNotification) {
@@ -93,7 +57,6 @@ class ViewController: UIViewController {
     
     // MARK: - UI Parts
     
-    @IBOutlet weak var tableView: UITableView!
     @IBOutlet weak var segmentedControl: UISegmentedControl!
     
     // MARK: - Action
@@ -108,32 +71,6 @@ class ViewController: UIViewController {
         if !tableView.decelerating {
             self.reloadAllPlugins()
         }
-        
-    }
-    
-    // MARK: - Table View Data Source
-
-    func tableView(tableView: UITableView, numberOfRowsInSection section: Int) -> Int
-    {
-        return Int(currentResult().count)
-    }
-    
-    func tableView(tableView: UITableView, cellForRowAtIndexPath indexPath: NSIndexPath) -> UITableViewCell
-    {
-        let plugin = currentResult()[UInt(indexPath.row)] as Plugin
-        
-        var cell = tableView.dequeueReusableCellWithIdentifier("Cell") as PluginTableViewCell
-        cell.rankingLabel.text = "\(indexPath.row + 1)"
-        cell.titleLabel.text = plugin.name
-        cell.noteLabel.text = plugin.note
-        cell.avaterImageView.sd_setImageWithURL(NSURL(string: plugin.avaterUrl))
-        
-        var formatter = NSDateFormatter()
-        formatter.dateFormat = "MM/dd/yy"
-                
-        cell.statusLabel.text = "\(Modes.Popularity.toIcon()) \(plugin.scoreAsString()) \(Modes.Stars.toIcon()) \(plugin.starGazersCount) \(Modes.Update.toIcon()) \(formatter.stringFromDate(plugin.updatedAt)) \(Modes.New.toIcon()) \(formatter.stringFromDate(plugin.createdAt))"
-        
-        return cell
     }
     
     // MARK: - Sign in
@@ -177,16 +114,31 @@ class ViewController: UIViewController {
             })
         }
     }
-
+    
+    // MARK: - Table View Data Source
+    
+    override func tableView(tableView: UITableView, numberOfRowsInSection section: Int) -> Int
+    {
+        return Int(currentResult().count)
+    }
+    
+    override func tableView(tableView: UITableView, cellForRowAtIndexPath indexPath: NSIndexPath) -> UITableViewCell
+    {
+        let plugin = currentResult()[UInt(indexPath.row)] as Plugin
+        
+        var cell = tableView.dequeueReusableCellWithIdentifier(PluginCellReuseIdentifier) as PluginTableViewCell
+        configureCell(cell, plugin: plugin, indexPath: indexPath)
+        
+        return cell
+    }
+    
     // MARK: - TableView Delegate
     
-    func tableView(tableView: UITableView, didSelectRowAtIndexPath indexPath: NSIndexPath) {
+    override func tableView(tableView: UITableView, didSelectRowAtIndexPath indexPath: NSIndexPath) {
         tableView.deselectRowAtIndexPath(indexPath, animated: true)
         
         let selectedPlugin = currentResult()[UInt(indexPath.row)] as Plugin
-        
-        var webViewController = M2DWebViewController(URL: NSURL(string: selectedPlugin.url), type: M2DWebViewType.AutoSelect)
-        navigationController?.pushViewController(webViewController, animated: true)
+        pushWebViewController(selectedPlugin.url)
     }
     
     // MARK: - Error
@@ -196,5 +148,5 @@ class ViewController: UIViewController {
         alert.addAction(UIAlertAction(title: "OK", style: UIAlertActionStyle.Default, handler: nil))
         self.presentViewController(alert, animated: true, completion: nil)
     }
-
+    
 }
