+++ AlcatrazTour/PluginTableViewCell.swift
@@ -2,6 +2,7 @@
 //  PluginTableViewCell.swift
 //  AlcatrazTour
 //
+//  Created by haranicle on 2015/03/14.
 //  Copyright (c) 2015年 haranicle. All rights reserved.
 //
 
@@ -14,5 +15,15 @@ class PluginTableViewCell: UITableViewCell {
     @IBOutlet weak var titleLabel: UILabel!
     @IBOutlet weak var noteLabel: UILabel!
     @IBOutlet weak var statusLabel: UILabel!
+
+    override func awakeFromNib() {
+        super.awakeFromNib()
+        // Initialization code
+    }
+
+    override func setSelected(selected: Bool, animated: Bool) {
+        super.setSelected(selected, animated: animated)
+
+    }
     
 }
